requirements.txt
playwright>=1.40.0
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0
zhipuai>=2.0.0
Pillow>=10.0.0